Moved tasks to better packages, reduced visibility on util class
