+++ src/main/groovy/org/shipkit/internal/gradle/git/tasks/GitPullTask.java
@@ -1,4 +1,4 @@
-package org.shipkit.internal.gradle.git;
+package org.shipkit.internal.gradle.git.tasks;
 
 import org.gradle.api.DefaultTask;
 import org.gradle.api.tasks.Input;
