+++ src/test/groovy/org/shipkit/internal/gradle/git/tasks/GitCommitTaskTest.groovy
@@ -1,4 +1,4 @@
-package org.shipkit.internal.gradle.git
+package org.shipkit.internal.gradle.git.tasks
 
 import org.gradle.testfixtures.ProjectBuilder
 import org.shipkit.gradle.git.GitCommitTask
